Moves the dummy memory used function.
